requirements.txt
Django==1.7
Scrapy==0.24
git+https://github.com/svven/tweepy.git#egg=tweepy
pywb==0.6.3
warcprox==1.3
uwsgi==2.0.6
django-celery==3.1.16
django-user-accounts==1.0
django-jsonfield==0.9.13
django-eztables==0.3.2
kombu==3.0.23
pinax-theme-bootstrap==5.5.1
requests==2.4.3
selenium==2.44.0
readability-lxml == 0.3.0.5
